+++ src/aceinna/devices/parsers/open_message_parser.py
@@ -49,7 +49,7 @@ class UartMessageParser(EventBase):
                     # find a whole frame
                     # self._parse_frame(self.frame, self.payload_len)
                     self._parse_message(
-                        packet_type, self.payload_len, self.frame[5:self.payload_len+5])
+                        packet_type, self.payload_len, self.frame)
 
                     self.find_header = False
                     self.payload_len = 0
@@ -77,16 +77,17 @@ class UartMessageParser(EventBase):
             'raw': raw_command
         }
 
-    def _parse_message(self, packet_type, payload_len, payload):
+    def _parse_message(self, packet_type, payload_len, frame):
+        payload = frame[5:payload_len+5]
         # parse interactive commands
         is_interactive_cmd = INPUT_PACKETS.__contains__(packet_type)
         if is_interactive_cmd:
-            self._parse_input_packet(packet_type, payload_len, payload)
+            self._parse_input_packet(packet_type, payload, frame)
         else:
             # consider as output packet, parse output Messages
-            self._parse_output_packet(packet_type, payload_len, payload)
+            self._parse_output_packet(packet_type, payload)
 
-    def _parse_input_packet(self, packet_type, payload_len, payload):
+    def _parse_input_packet(self, packet_type, payload, frame):
         payload_parser = match_command_handler(packet_type)
         if payload_parser:
             data, error = payload_parser(
@@ -95,12 +96,13 @@ class UartMessageParser(EventBase):
             self.emit('command',
                       packet_type=packet_type,
                       data=data,
-                      error=error)
+                      error=error,
+                      raw=frame)
         else:
             print('[Warning] Unsupported command {0}'.format(
                 packet_type.encode()))
 
-    def _parse_output_packet(self, packet_type, payload_len, payload):
+    def _parse_output_packet(self, packet_type, payload):
         # check if it is the valid out packet
         payload_parser = None
         is_other_output_packet = OTHER_OUTPUT_PACKETS.__contains__(packet_type)